+++ meta_package_manager/managers/apm.py
@@ -23,7 +23,7 @@ from typing import Iterator
 from click_extra.platform import LINUX, MACOS, WINDOWS
 
 from ..base import Package, PackageManager
-from ..capabilities import search_capabilities
+from ..capabilities import search_capabilities, version_not_implemented
 
 
 class APM(PackageManager):
@@ -420,7 +420,8 @@ class APM(PackageManager):
                     latest_version=pkg["version"],
                 )
 
-    def install(self, package_id: str) -> str:
+    @version_not_implemented
+    def install(self, package_id: str, version: str | None = None) -> str:
         """Install one package.
 
         .. code-block:: shell-session