+++ meta_package_manager/managers/chocolatey.py
@@ -23,6 +23,7 @@ from typing import Iterator
 from click_extra.platform import WINDOWS
 
 from ..base import Package, PackageManager
+from ..capabilities import version_not_implemented
 
 
 class Choco(PackageManager):
@@ -142,7 +143,8 @@ class Choco(PackageManager):
         for package_id, latest_version in regexp.findall(output):
             yield self.package(id=package_id, latest_version=latest_version)
 
-    def install(self, package_id: str) -> str:
+    @version_not_implemented
+    def install(self, package_id: str, version: str | None = None) -> str:
         """Install one package.
 
         .. code-block:: shell-session