+++ meta_package_manager/managers/emerge.py
@@ -23,6 +23,7 @@ from typing import Iterator
 from click_extra.platform import LINUX
 
 from ..base import Package, PackageManager
+from ..capabilities import version_not_implemented
 
 
 class Emerge(PackageManager):
@@ -220,7 +221,8 @@ class Emerge(PackageManager):
                 latest_version=version,
             )
 
-    def install(self, package_id: str) -> str:
+    @version_not_implemented
+    def install(self, package_id: str, version: str | None = None) -> str:
         """Install one package.
 
         .. code-block:: shell-session