+++ meta_package_manager/managers/flatpak.py
@@ -23,7 +23,7 @@ from typing import Iterator
 from click_extra.platform import LINUX
 
 from ..base import Package, PackageManager
-from ..capabilities import search_capabilities
+from ..capabilities import search_capabilities, version_not_implemented
 
 
 class Flatpak(PackageManager):
@@ -169,7 +169,8 @@ class Flatpak(PackageManager):
                 latest_version=version,
             )
 
-    def install(self, package_id: str) -> str:
+    @version_not_implemented
+    def install(self, package_id: str, version: str | None = None) -> str:
         """Install one package.
 
         .. code-block:: shell-session