+++ meta_package_manager/managers/npm.py
@@ -24,7 +24,7 @@ from boltons.iterutils import remap
 from click_extra.platform import LINUX, MACOS, WINDOWS
 
 from ..base import Arg, NestedArgs, Package, PackageManager
-from ..capabilities import search_capabilities
+from ..capabilities import search_capabilities, version_not_implemented
 
 
 class NPM(PackageManager):
@@ -277,7 +277,8 @@ class NPM(PackageManager):
                     latest_version=pkg.get("version"),
                 )
 
-    def install(self, package_id: str) -> str:
+    @version_not_implemented
+    def install(self, package_id: str, version: str | None = None) -> str:
         """Install one package.
 
         .. code-block:: shell-session