+++ meta_package_manager/managers/yarn.py
@@ -30,7 +30,7 @@ else:
 from click_extra.platform import LINUX, MACOS, WINDOWS
 
 from ..base import Package, PackageManager
-from ..capabilities import search_capabilities
+from ..capabilities import search_capabilities, version_not_implemented
 
 
 class Yarn(PackageManager):
@@ -268,7 +268,8 @@ class Yarn(PackageManager):
                     latest_version=package["version"],
                 )
 
-    def install(self, package_id: str) -> str:
+    @version_not_implemented
+    def install(self, package_id: str, version: str | None = None) -> str:
         """Install one package.
 
         .. code-block:: shell-session